requirements.txt
tensorflow==2.13.0
numpy==1.24.3
Pillow>=9.0.0
scikit-learn==1.3.0
matplotlib>=3.5.0
python-pptx>=0.6.21

